fusion: remove i2c_device_id parameter from probe() function
Initially, the i2c probe() function had two parameters: an i2c_client and an
i2c_device_id. A new function called probe_new() was introduced, which removed
the second parameter, and both functions co-existed in order to transition from
one to another. Eventually, the parameter was removed from probe() and so was
probe_new().

The fusion driver didn't make use of this parameter, so removing it isn't an
issue.

https://onedigi.atlassian.net/browse/DEL-9003

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>