rtc: stm32: write RTC calibration values from DT
If for whatever reason the input frequency of the oscillator is not
accurate, the RTC_CALR register can be used to add/subtract counts from
the input clock.

RTC_CALR register has two fields:
 - CALP (bit 15) must be set to increase the count
 - CALM (bits 8:0) must be set to a value to accurately decrease the count

To be able to set CALP, PREDIV_A must have a value > 3.

This commit creates a couple of DT properties to set the calibration
register:
 - digi,calib-calp (boolean)
 - digi,calib-calm (u32)

When CALP flag must be set, set PREDIV_A to highest value.

Calculation:
 - CALP = 1 if Frtclk < Fcal (otherwise =0)
 - CALM = [512*CALP*Frtclk + (Fcal-Frtclk)*(512*CALP - 2^20)]/Fcal
where
 * Frtclk is the actual input frequency
 * Fcal is the nominal frequency (32768)

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://onedigi.atlassian.net/browse/DEL-8987