ARM: dts: add overlay for hardware version 1 of CCMP15 SOM
Hardware version 1 (HV=1) of CCMP15 SOMs have a slightly big capacitance
on the LSE oscillator lines that results in a low input frequency
(average 32765.42Hz, or 78.65ppm).
As a consequence, the RTC count suffers an average drift of around -6.8
seconds/day.

Correct this with the RTC calibration register.

Calculation is as follows:
 - CALP = 1 if Frtclk < Fcal (otherwise =0)
 - CALM = [512*CALP*Fcrtclk + (Fcal-Frtclk)*(512*CALP - 2^20)]/Fcal
where
 * Frtclk is the input frequency
 * Fcal is 32768

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://onedigi.atlassian.net/browse/DEL-8987