mach-imx6ul: remove dead code
The xbee reset cannot be run at this state, as the gpio node is not probed
yet in the system, so you get an EPROBE_DEFER.

This error happened as well in previous kernel versions, but there was no
code there printing the failing situation.

The XBee Reset is done in the specific rootfs script that initializes the XBee
node.

https://onedigi.atlassian.net/browse/DEL-9019

Signed-off-by: Isaac Hermida <isaac.hermida@digi.com>