ARM64: dts: ccmp25-dvk: split Peripheral I2C bus on I2C1 and I2C2
This commit adds support to ConnectCore MP2 DVK v2, where the Peripheral I2C
bus was splited to use the I2C1 at 3V3 and the I2C2 at 1V8, moving the
peripherals accordingly. Also was removed the workaround to fix the HPD line
for the Lontium MIPI-DSI brigde due to it is fixed in DVK v2.

https://onedigi.atlassian.net/browse/DEL-8995

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>