Revert "drm/stm: dsi: fixate maximum lane rate to 1 Gbps"
This workaround introduces an intermittent issue on the ccmp25-dvk in which the
HDMI image trembles constantly depending on the contents of the output. Also,
this workaround was only modifying the lane_mbps variable internal to the STM
MIPI-DSI driver, while leaving the variable used in the parent driver
(drivers/gpu/drm/bridge/synopsys/dw-mipi-dsi.c) untouched, causing a slight
mismatch when calculating timings.

This reverts commit 4403ccc623cf01e30334296c668884952151ee39.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>