drm/stm: dsi: make PLL freq increase percentage configurable via device tree
Even though a 30% increase in the PLL frequency has been proven to work on the
ccmp25-dvk with several HDMI monitors, we can't guarantee that it'll work in
every case. Add the possibility to configure this percentage via an optional
device tree parameter "digi,burst-pll-inc-pcnt", which allows values between 0%
and 50%. Without this parameter, 30% is used by default.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>