drm/stm: dsi: increase PLL freq by 30% instead of 20% when using burst mode
While it's true that the PLL frequency must be increased when burst mode is
enabled, the hardcoded increment of 20% causes flickering to happen when using
an HDMI display in FullHD on the ccmp25-dvk. After testing several values on
different monitors, an increment of 30% results in the most stable output, so
use this value instead.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>