ARM64: stm32mp25: disable USB OHCI/EHCI nodes on dtsi
The stm32mp251.dtsi include file enables both OHCI and
EHCI subnodes of the usb2h host controller.
Then the board device trees disable one of them (the OHCI)
so that the host controller works in high-speed mode (EHCI).

The correct way should be the opposite:
- the SoC include file disables both subnodes
- the board files enable the subnode they want (EHCI by default)

This commit does exactly this for all existing boards (STM and Digi).
The commit doesn't make any functional change, but it's more
readable to see the nodes that we enable than presume that EHCI
is enabled by disabling OHCI.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>