arm64: dts: ccimx91: fix reserved memory ranges
The ccimx91 is a 512MB memory device. The memory ranges copied from the
ccimx93 are not valid. Reduce the linux CMA to 96MB and move the ELE
reserved memory within the 512MB range (just above the Optee).

Also, disable the thermal support, as it is not working in this version
of the kernel for this platform.

https://onedigi.atlassian.net/browse/DEL-9171

Signed-off-by: Javier Viguera <javier.viguera@digi.com>