arch: arm64: ccmp2_defconfig: disable NF_TABLES
In terms of CVE patching, this configuration option tends to add a lot of
overhead, and we don't enable it on any other platform. For starters, there is
a significant amount of CVEs that affect this option, and on top of that, the
fixes for said CVEs usually involve several commits in order to ensure that the
patch context is exact.

All in all, we should remove this option from our default configuration to
avoid this overhead.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>