arch: arm64: ccmp2_defconfig: do an in-depth cleanup of the defconfig
The defconfig was enabling a lot of unnecessary options, especially modules,
resulting in bloated compilation times and image sizes. Using the
ccimx9_defconfig from our v6.1/nxp/dey-4.0/maint branch as reference, disable
several options and configure modules as built-in when possible.

When it comes to STM32-specific configurations, maintain a conservative
approach. We've observed that changing some of these options from modules to
built-in (or viceversa) can lead to unexpected behavior, including system
crashes. For now, leave these options as they are.

After this change:

    * Compilation time (kernel + dtbs + modules) is reduced by nearly 50%
      (on my build machine, from 4m3s to 2m13s)
    * Compressed Image.gz size is reduced by 296 KiB, uncompressed Image size
      is reduced by over 2 MiB
    * Number of modules is reduced from 884 (185 MiB) to 100 (12 MiB). This
      directly affects DEY image sizes (recovery image is reduced by 11 MiB,
      dey-image-lvgl rootfs image is reduced by 48 MiB)

https://onedigi.atlassian.net/browse/DEL-9189

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>