arm64: dts: ccimx9: make CAN stby regulators depend on reg_5v
The 5v regulator powers the CAN transceivers, so when the regulator is
off, CAN fails. On the ccimx93, the regulator was powered due to other
peripherals (HDMI bridge), so CAN was working. On the other hand, the
ccimx91 does not have those peripherals, the 5v regulator is off, and
thus CAN fails.

Configure the CAN standby regulators properly as dependents of the 5v
regulator so it is powered on when CAN is used.

https://onedigi.atlassian.net/browse/DEL-9209

Signed-off-by: Javier Viguera <javier.viguera@digi.com>