drivers: ov5640: increase reset time to >10ms
According to the OV5640 sensor datasheet, it requires, at least,
a 1ms delay after asserting the reset line for it to stabilize.

Setting a stabilization time between 1 and 2ms can be too close to the
minimum requirement, and depending on the accuracy of the processor,
it can fall short.

Increase the minimum reset time to be at least 10ms so we have plenty of
margin and the minimum time is assured.

Originally fixed on 74b7b959fe77d3e98d1753ad279d581cc75c8fa2 on a
CC8M Nano.

Tested on a CCMP25 DVK.

https://onedigi.atlassian.net/browse/DEL-9187

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>