ARM64: dts: ccmp25: bluetooth: fix power regulator entry
Currently, the bluetooth interface is attached and managed by the HCI_UART
driver, which does not have support to manage the power supply.
So, just keep enable that regulator.
This will need to be modified when the support is migrated to the bcm43438
driver, as the driver will manage that power supply.

https://onedigi.atlassian.net/browse/DEL-9248

Signed-off-by: Isaac Hermida <isaac.hermida@digi.com>