ARM64: dts: ccmp25-mfg: enable vddio1 regulator as always-on
This is a regulator internal to the SoC, and it powers lines PE0 through
PE5, which are used on wireless variants as the wireless chip SDIO port.

On non-wireless variants, these lines are available externally, but can
only be used when the vddio1 regulator is enabled, even as GPIOs.

Enable this regulator by default as always-on, to ensure these lines are
available and can be tested through GPIO loopbacks.

https://onedigi.atlassian.net/browse/DEL-9268

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>