ccimx6ul_defconfig: configure cfg80211 as module
The cfg80211 support relies on a firmware file (regulatory.db).
When cfg80211 is built into the kernel (as a built-in), loading the certificate
for the regulatory database fails with the following error:

"""
cfg80211: Loading compiled-in X.509 certificates for regulatory database
Loaded X.509 cert 'sforshee: 00b28ddf47aef9cea7'
Loaded X.509 cert 'wens: 61c038651aabdcf94bd0ac7ff06c7248db18c600'
platform regulatory.0: Direct firmware load for regulatory.db failed with error -2
"""

Error -2 means "File not found". When cfg80211 is built directly into
the kernel, it tries to load the regulatory.db file during early boot.
At this point, the root filesystem is not fully mounted yet,
making /lib/firmware/regulatory.db inaccessible.

In earlier versions, this error could occur as well, but it was resolved
because CRDA was installed, which ultimately handled the loading of the firmware.

The simplest and most effective solution is to configure CFG80211 as a module.
Additionally, CRDA support is no longer needed.

https://onedigi.atlassian.net/browse/DEL-9402

Signed-off-by: Isaac Hermida <isaac.hermida@digi.com>