brcmfmac: fix channel issue after country code changed
There is a proprietary patch in RPi which isn't patched by us, so
customer can change country code by iw or wpa_supplicant.conf in
previous release.

MR601 had committed from Indrik release which caused this feature
failing, it is not a official flow of changing country code of the
FMAC is allowable. So it's not a regression.

[Analysis]
In brcmf_fill_band_with_default_chanlist(), we expected to get a
default channel list, but existing channel list will not be updated due
to a condition check.

[Solution]
Reset band and channel list at the beginning.

Fixes SWLINUX-4931

Signed-off-by: Owen Huang <Owen.Huang@infineon.com>