imx6ul: avoid lowering ddr3 freq for cc6ul
By default, the DDR refresh mechanism sources from the internal
ring oscillator or the external 32kHz RTC_XTAL.
The CC6UL does not have a crystal, so the internal ring oscillator
is used. This, however, is sensitive to noise and temperature
and under certain conditions may result in extra refresh cycles.
To avoid these, the MMDC has been configured by the boot loader
to source from the DDR clock instead.
To guarantee the correct refresh rate, avoid changing the DDR
frequency for the CC6UL.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://onedigi.atlassian.net/browse/DEL-9264