ARM: dts: ccmp13: declare vrefbuf, without consumer
Enabling the internal voltage regulator (VREFBUF) in OPTEE
doesn't have any effect, as long as there is no consumer
of it in the Linux device tree.
To facilitate using VREFBUF, this regulator has been enabled
in OPTEE, so we can define it sibiling through SCMI in the
Linux device tree.
By default, it has no consumer, so it remains disabled and
doesn't output its voltage at VREF+ pad.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://onedigi.atlassian.net/browse/DEL-9436