pwm: pwm-mca: fix pwm configuration through sysfs
Current implementation was only enabling the pwm if a second configuration of
the duty cycle was done while the pwm was enabled.

Simplified current vs desired state logic.

The apix-pwm example was also failing in the first configuration, but seemed to
work since it changes the duty cycle every 150mS.

https://onedigi.atlassian.net/browse/DEL-9495

Signed-off-by: Hector Bujanda <hector.bujanda@digi.com>