[PATCH] [MICRSE-3165] RNG workaround for i.MX6UL
Implement RNG workaround for i.MX6UL

The i.MX6UL processor, under some conditions, could fail the validation
of the entropy generated by the TRNG. This is caused by a problem on the
RNG health test. The workaround performs the validation using a software
routine, running from the OCRAM, instead of using the HW support. The
bootloader loads the firmware to do the validation in the OCRAM memory
if the RNG workaround is enabled and, when the error condition is
detected, performs the validation using the software routine loaded in
the OCRAM.

https://onedigi.atlassian.net/browse/DEL-8673

Signed-off-by: yanxing cui <yanxing.cui@nxp.com>
Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>