ARM: dts: ccmp13-dvk: add 5V board regulator
The 5V load switch on the DVK (activated by a GPIO) was not defined.
This is powering the CAN transceiver on the DVK. On the ccmp15, this
regulator is defined since it also powers other interfaces (display).

Since the CAN driver doesn't support regulators we need to set this as
an always-on regulator, so that the CAN transceiver works out-of-box.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://onedigi.atlassian.net/browse/DEL-9637