ARM: dts: ccmp13-dvk: fix power key button compatible string
The power key device driver for ConnectCore modules was recently generalized to
support all ConnectCore MP platforms under Linux kernel v6.x. As part of this
update, the compatible string was modified to reflect the new unified driver.

However, support for the ccmp13 platform was inadvertently missed during the
migration. This commit fixes the compatible string in the ccmp13 device tree to
ensure proper functionality with the updated driver.

https://onedigi.atlassian.net/browse/DEL-9628

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>