ARM: dts: ccmp15/13: add cpu_thermal and trip_points
On the CCMP13, the thermal entry with cooling map that exists in
stm32mp13xd.dtsi is for version D of the CPU which supports different CPU
frequencies. The CCMP13 assembles flavor C of the CPU that only supports
650MHz frequency, so the CPU is not a valid cooling device.

Remove the include and replicate the cpu_thermal node without cooling map
as reference for other potential cooling devices and specially to establish
a critical point to guard the rest of the SOM components as per thermal
experiments carried out.

https://onedigi.atlassian.net/browse/DEL-9469

Signed-off-by: Hector Palacios <hector.palacios@digi.com>