ARM: dts: ccmp15: update CCMP15 DVKv1 overlay for HPD with LT8912B HDMI bridge
This commit updates the Device Tree overlay for the CCMP15 DVKv1 to support Hot
Plug Detect (HPD) with the Lontium LT8912B HDMI bridge using the upstream
driver. The update aligns the overlay with recent changes introduced in the
CCMP1 platform, where the upstream LT8912B driver is now in use.

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>