adc: cc6ul: remove Digi-specific DT properties 'max-channel' and 'adc-ch-list'
This effectively resets the driver to NXP BSP status.
These properties were added to workaround a limitation in old v4.14 driver
that was fixed long ago in NXP BSP.

Without the properties, the ADC now exposes all ADC channels. It is still
a responsibility of the user's DT to configure the correct pinctrl for
the pads to behave as ADC.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://onedigi.atlassian.net/browse/DEL-9763
https://onedigi.atlassian.net/browse/DEL-6469