ARM: mach-imx6q: implement error handling for digi,pwrdown GPIO
Commit dff2b52198e7f89feee7a8fb0a5a5e77f4120921 ("arch: imx6q: migrate from
deprecated GPIO API to gpiod one") introduced incorrect handling of the GPIO
descriptor. When fwnode_gpiod_get_index() returns an error, the code attempts
to use the error pointer as a valid GPIO descriptor, leading to unpredictable
behavior. This commit adds proper error handling to skip the use of the GPIO if
the descriptor retrieval fails.

https://onedigi.atlassian.net/browse/DEL-9759

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>