ARM: dts: imx6qdl/6qp: define Wi-Fi enable via MMC power sequence provider
This commit configures the power sequence for the wireless variants of the
i.MX6QDL and i.MX6QP platforms by defining the control of the wireless enable
line through the MMC power sequence provider.

The update ensures that the Wi-Fi chip is properly powered using the MMC
framework, aligning with recent changes that remove platform-specific power
control logic and instead rely on device tree-managed regulators.

https://onedigi.atlassian.net/browse/DEL-9759

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>