Bluetooth: hci_event: accept legacy reply for LE Read Supported States
Some legacy Bluetooth controllers (e.g., AR6223) respond to the LE Read Supported
States HCI command (0x201c) with only 5 bytes instead of the expected 9 bytes
(1 byte for status + 8 bytes for the LE feature mask). This causes the host to
reject the event due to a mismatch in the expected completion length:

    Bluetooth: hci0: unexpected cc 0x201c length: 5 < 9
    Bluetooth: hci0: Opcode 0x201c failed: -38

Ideally, this behavior should be managed by the HCI_QUIRK_VALID_LE_STATES quirk
flag. However, this lenght check is currently enforced unconditionally, and the
quirk is not evaluated in this element. This commit relaxes the length check
specifically for this command, allowing shorter replies to be accepted. This
ensures compatibility with legacy controllers that don't adhere to the current
specification but are otherwise functional.

https://onedigi.atlassian.net/browse/DEL-9759

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>