ARM64: dts: ccimx95: add support to PCIe ports
PCIe0 port is routed to miniPCIe connector.
Select-able 0-ohm resistors can swap the PCIe0 port with the PCIe1 port.
No RESET or CLK_REQ_B lines are routed to the connector.
Power (3.3V) comes from the 3v3 DVK regulator and a 1.5V LDO on the DVK
that is not controlled by any GPIO.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>