ARM64: dts: ccimx95: add SCMI PMIC regulators
Regulators from the PMIC are originally managed by the System Manager.
The SM now exposes a few regulators that are used outside of the SOM:

- VDD_3V3 (SW1)
- VDD_1V8 (SW3)
- LDO1 (VDD_LDO1)
- LDO2 (VDD_SDIO2)

By declaring them on the device tree, we can link them to drivers and
let Linux manage them.

Note: On the first prototypes LDO1 works as a 3V3 load switch instead
of as an LDO, so it won't be used in these prototypes.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>