arm64: defconfig: set stm32 rtc as built-in in addons fragment
The stm32 rtc is a clock provider for the node wifi-pwrseq.

Set the driver as built-in to prevent the probe of the node
wifi-pwrseq to be deferred while waiting for the availability
of the root file system.

Signed-off-by: Antonio Borneo <antonio.borneo@foss.st.com>
Change-Id: I09ec3b9dc153a34289d064770d6cc7f87cd1a422
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/503624
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Amelie DELAUNAY <amelie.delaunay@foss.st.com>
Domain-Review: Amelie DELAUNAY <amelie.delaunay@foss.st.com>