ARM64: dts: ccimx95: substitute PMIC VDD_3V3 and 1V8 with dummy regulators
There doesn't seem to be a clear registered handler for reboot or shutdown
path from Linux to SM.
When a 'reboot' is requested, Linux falls back to run the global
device_shutdown() path.
The SM will turn these regulators off in their shutdown path or via a
global “disable regulators on reboot” policy. The current SM firmware
doesn't get to know if Linux has marked these regulators as "always-on".

Powering these regulators off removes power to the SOM who is then not
able to continue the reset (it is not powered).

By removing the PMIC regulators, we fully rely on SM managing them
though apparently support is incomplete (for instance they are not powered
off on the poweroff path from Linux), so this may need to be revisited in
future BSP upgrades.

Create dummy regulators to satisfy the supplies of other devices.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://onedigi.atlassian.net/browse/DEL-9804