usb: dwc2: put the core in power saving when no enumeration
When the enumeration doesn't complete, the suspend irq lets the
core active (ignore suspend request during enumeration).
The enumeration normally ends at some point, within connect timings
range. Allow to put the core in one of the power saving modes, once
passed the timeout.

A typical case is the disconnect of a cable. The controller, gets a
suspend interrupt, which is ignored before enumeration. Until it
gets plugged to another (A or B) device, this timeout allows to
reach low power state.

Another case is: the user configures the gadget driver, but the cable
isn't plugged to a host. To avoid letting the controller active while
not plugged, use the same timeout in dwc2_hsotg_core_connect.

The timeout gets canceled, when the enumeration completes.

Signed-off-by: Fabrice Gasnier <fabrice.gasnier@foss.st.com>
Change-Id: Ifdedc2458198a77379c67a7956ac6c58e049d771
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/512990
Reviewed-by: Amelie DELAUNAY <amelie.delaunay@foss.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>