ARM64: dts: ccimx95: disable LVDS1 when LVDS0 is selected
Both LVDS could work at the same time but since their touch controllers
are connected to the same I2C bus with the same I2C address, it would
generate conflicts and would not work for any of the panels.

To avoid issues, when LVDS0 overlay is applied, ensure LVDS1 channel is
disabled, in the same way that LVDS1 overlay already is disabling LVDS0
channel.

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>